Pequeña Refactorización en Vista Se han separado los botones del flujo principal. Aunque parezca pequeño el cambio servirá de mucho para cuando se agregen nuevas funcionalidades.
