+++ func/principal.js
@@ -1,7 +1,7 @@
 import { configurar_idioma } from "./config/idioma.js";
 import { configurar_tema } from "./config/tema.js";
 import { cargar_editor } from "./modo/editor/editor.js";
-import { cargar_vista } from "./modo/vista.js";
+import { cargar_vista } from "./modo/vista/vista.js";
 import { cargar_tablero } from "./modo/tablero/tablero.js";
 import { configurar_cabecera } from "./util/cabecera.js";
 import { cargar_formularios } from "./ventanas/formulario.js";
